apps/ai-service/requirements.txt
# FastAPI & Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LLM Providers
google-generativeai>=0.3.0
openai>=1.3.0
anthropic>=0.7.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.25.0
Pillow>=10.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
